+++ tests/testthat/test-shinytest2-teal_slices.R
@@ -26,7 +26,7 @@ testthat::test_that("e2e: teal_slices filters are initialized when global filter
     c("4", "6")
   )
   testthat::expect_identical(
-    app$get_active_filter_selection("mtcars", "drat", is_numeric = TRUE),
+    app$get_active_filter_selection("mtcars", "drat"),
     c(3, 4)
   )
   testthat::expect_identical(
@@ -67,7 +67,7 @@ testthat::test_that("e2e: teal_slices filters are initialized when module specif
     app$get_active_filter_selection("mtcars", "cyl"),
     c("4", "6")
   )
-  testthat::expect_null(app$get_active_filter_selection("mtcars", "drat", is_numeric = TRUE))
+  testthat::expect_null(app$get_active_filter_selection("mtcars", "drat"))
   testthat::expect_null(app$get_active_filter_selection("mtcars", "gear"))
 
   app$navigate_teal_tab("Module_2")
@@ -80,7 +80,7 @@ testthat::test_that("e2e: teal_slices filters are initialized when module specif
     c("setosa", "versicolor", "virginica")
   )
   testthat::expect_identical(
-    app$get_active_filter_selection("mtcars", "drat", is_numeric = TRUE),
+    app$get_active_filter_selection("mtcars", "drat"),
     c(3, 4)
   )
   testthat::expect_identical(
@@ -89,13 +89,5 @@ testthat::test_that("e2e: teal_slices filters are initialized when module specif
   )
   testthat::expect_null(app$get_active_filter_selection("mtcars", "cyl"))
 
-  app$set_active_filter_selection("iris", "Species", "setosa")
-  app$navigate_teal_tab("Module_1")
-  app$wait_for_idle(timeout = default_idle_timeout)
-
-  testthat::expect_identical(
-    app$get_active_filter_selection("iris", "Species"),
-    "setosa"
-  )
   app$stop()
 })